+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- [TD-8083] Add user group management features
+
 ## [8.3.0] 2026-03-12
 
 ### Changed